Put KeyMappingX11 stuff inside its own scope
This avoids collisions with other "concurrent" key mappers.
